+++ react-rest-countries/src/components/Country/Country.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Country = () => {
+    return (
+        <div>
+            <h2>Countries in other file</h2>
+        </div>
+    );
+};
+
+export default Country;
