+++ react-rest-countries/src/components/Header/Header.css
@@ -1,0 +1,23 @@
+.header {
+    display: flex;
+    justify-content: space-around;
+
+}
+.logo a {
+    text-decoration: none;
+    margin-top: 0;
+}
+nav ul {
+    display: flex
+    
+}
+nav ul li {
+    list-style: none;
+    margin: 8px 30px 0 0;
+}
+nav ul li a {
+    text-decoration: none;
+    margin-right: 40px;
+    font-size: 18px;
+    font-weight: bold;
+}
